refactor: 웹 구동을 위한 코드 분리
